requirements.txt
gym~=0.17.2
numpy~=1.19.0
requests~=2.24.0
scipy~=1.4.1
pyglet~=1.5.0
matplotlib~=3.2.2
pip~=20.1.1
wheel~=0.34.2
rsa~=4.6
pyasn1~=0.4.8
six~=1.15.0
gast~=0.3.3
h5py~=2.10.0
future~=0.18.2
setuptools~=49.2.0
cachetools~=4.1.1
certifi~=2020.6.20
urllib3~=1.25.9
chardet~=3.0.4
python-dateutil~=2.8.1
Markdown~=3.2.2
oauthlib~=3.1.0
idna~=2.10
Werkzeug~=1.0.1
cycler~=0.10.0
pyparsing~=2.4.7
kiwisolver~=1.2.0
tensorflow~=2.2.0
tensorboard~=2.2.2
wrapt~=1.12.1
astunparse~=1.6.3
termcolor~=1.1.0
cloudpickle~=1.3.0
kafka-python~=2.0.1